Creates util for modules and adds method get_room_from_jid in it.
